semiautonomous-agents/docparse/extractor/pyproject.toml
[project]
name = "docparse"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "cloudevents>=2.0.0",
    "fastapi>=0.136.1",
    "google-auth>=2.49.2",
    "google-cloud-storage>=3.10.1",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "pillow>=12.2.0",
    "pydantic>=2.13.3",
    "pypdfium2>=5.7.1",
    "rich>=15.0.0",
    "typer>=0.24.2",
    "uvicorn[standard]>=0.46.0",
]

[project.scripts]
docparse = "docparse:main"
docparse-service = "docparse.service:app"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"
